+++ Changelog.md
@@ -4,6 +4,8 @@ Features:
  * Inline Assembly: Issue warning for using jump labels (already existed for jump instructions).
 
 Bugfixes:
+ * Type Checker: Suggest the experimental ABI encoder if using ``struct``s as function parameters
+   (instead of an internal compiler error).
 
 ### 0.4.19 (2017-11-30)